backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
httpx==0.28.0
spotipy==2.24.0
mutagen==1.47.0
pyacoustid==1.3.0
musicbrainzngs==0.7.1
aiofiles==24.1.0
python-multipart==0.0.17
websockets==14.1
python-dotenv==1.0.1
Pillow==11.0.0
requests==2.32.3
